Remove typeDefs and resolvers from graphql server file
