+++ server/graphql/resolvers/index.js
@@ -1,0 +1,8 @@
+const noticeResolvers = require('./noticeResolver');
+const userResolvers = require('./userResolver');
+
+module.exports = {
+  Query: {
+    ...noticeResolvers.Query
+  }
+}
