+++ server/graphql/typeDefs.js
@@ -1,0 +1,15 @@
+const {gql} = require('apollo-server');
+
+module.exports = gql`
+  type Notices{
+    id: ID
+    name: String
+    user: String
+    platformId: String
+    createdAt: String
+
+  }
+  type Query{
+    fetchAllNotices: [Notices]
+  }
+`
